% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_descr_lanuv}
\alias{check_descr_lanuv}
\title{Check and set the default description file for LANUV data}
\usage{
check_descr_lanuv(descr)
}
\arguments{
\item{descr}{Character. For example, `"pegel_stand"`, `"pegel_stationen"`,
`"hydrmet_N"`, `"hydrmet_tag"`, or a custom file name.}
}
\value{
Character string with the description file name.
}
\description{
This function checks whether the input argument `descr` is one of the
predefined default descriptions or a user-defined file. If valid, the
corresponding file name is returned.
}
\examples{
check_descr_lanuv("pegel_stand")
check_descr_lanuv("my_file.txt")
}
