% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_hub}
\alias{check_hub}
\title{Check and set the data hub}
\usage{
check_hub(hub)
}
\arguments{
\item{hub}{Character. Either a keyword for the predefined hubs
(`"kisters"`, `"raw_nrw"`, `"verified_level_nrw"`,
`"verified_runoff_nrw"`) or a custom URL.}
}
\value{
Character string containing the API URL.
}
\description{
This function checks whether the input argument `hub` is one of the
predefined default hubs or a user-defined URL. If valid, the corresponding
API URL is returned.
}
\examples{
check_hub("kisters")
check_hub("https://example.com/api/")
}
