% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_query_lanuv}
\alias{check_query_lanuv}
\title{Check and set the default query for LANUV data}
\usage{
check_query_lanuv(query)
}
\arguments{
\item{query}{Character. Either `"pegel"`, `"hydromet"`, or a custom
file name.}
}
\value{
Character string with the query file name.
}
\description{
This function checks whether the input argument `query` is one of the
predefined default queries or a user-defined file. If valid, the
corresponding file name is returned.
}
\examples{
check_query_lanuv("pegel")
check_query_lanuv("custom_file.zip")
}
