% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{descr_names}
\alias{descr_names}
\title{Extract file names from ZIP archives on a web page}
\usage{
descr_names(page_url, zip_names)
}
\arguments{
\item{page_url}{Character. Base URL of the web page containing the ZIP files.}

\item{zip_names}{Character vector. Names of the ZIP files.}
}
\value{
List with ZIP file names as keys and vectors of contained files as values.
}
\description{
Downloads multiple ZIP files from the web page and extracts the contained file names.
}
