% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{download_hydrodata_nrw}
\alias{download_hydrodata_nrw}
\title{Download hydro data from ZIP archive and read shapefile or CSV}
\usage{
download_hydrodata_nrw(hub, query, descr = NULL)
}
\arguments{
\item{hub}{Character. Base URL or path.}

\item{query}{Character. File path or query string.}

\item{descr}{Character or NULL. Optional: Only extract these files.}
}
\value{
An sf object (for shapefile), data.frame (for CSV), or NULL if no file found.
}
\description{
Downloads a ZIP file from `hub + query`, extracts either all files or only
the specified `descr`, and reads a shapefile or CSV file.
}
