% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{list_page_zips}
\alias{list_page_zips}
\title{List ZIP file names from a web page}
\usage{
list_page_zips(page_url)
}
\arguments{
\item{page_url}{Character. URL of the web page.}
}
\value{
Character vector with ZIP file names.
}
\description{
Loads the HTML page and extracts all ZIP file names from JSON-like structures.
}
