% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{download_hydrodata_nrw}
\alias{download_hydrodata_nrw}
\title{Download hydro data from ZIP archive and read shapefile or CSV}
\usage{
download_hydrodata_nrw(hub, query, descr = NULL)
}
\arguments{
\item{hub}{Character. Base URL.}

\item{query}{Character. File path or query string.}

\item{descr}{Character or NULL. Optional: restrict extraction to these files.}
}
\value{
An `sf` object (for shapefile), a `tibble` (for CSV), or `NULL` if no suitable file is found.
}
\description{
Downloads a ZIP archive from `hub + query`, extracts either all files or
the specified `descr`, and loads shapefiles as `sf` objects or CSVs as tibbles.
}
\details{
ZIP files are downloaded and removed after extraction.
}
\examples{
\dontrun{
shp <- download_hydrodata_nrw("verified_level_nrw", "stations.zip")
}
}
