% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{get_pagetree}
\alias{get_pagetree}
\title{Create metadata table with information about ZIP files and their contents}
\usage{
get_pagetree(page_url)
}
\arguments{
\item{page_url}{Character. URL of the page with ZIP files.}
}
\value{
data.frame with metadata about ZIP and contained files.
}
\description{
Reads all ZIP file names from a page and extracts information from the file names of the ZIP contents.
}
