% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{get_rawdata_nrw}
\alias{get_rawdata_nrw}
\title{Download and read raw NRW data from ZIP file}
\usage{
get_rawdata_nrw(hub, query, descr)
}
\arguments{
\item{hub}{Character. Base URL or path to the data hub.}

\item{query}{Character. Path or query string appended to `hub`.}

\item{descr}{Character. Filename inside the ZIP file to extract and read.}
}
\value{
A `data.frame` or `tibble` with the loaded data.
}
\description{
This function downloads a ZIP file from `hub + query`,
extracts a file named `descr`, and reads this CSV file.
}
