% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{get_rawdata_nrw}
\alias{get_rawdata_nrw}
\title{Get the raw data from the LANUV datasource}
\usage{
get_rawdata_nrw(hub, query, descr)
}
\arguments{
\item{hub}{The file location you are querying data from. Either one of the defaults or a URL.
See \href{https://github.com/iso-code/interfaces}{README}.}

\item{query}{the type of data either precipitation or water level data.}
}
\value{
A dataframe with three columns: station_no, time and value.
}
\description{
Returns a dataframe containing the most recent data.
}
\examples{
\dontrun{
get_rawdata_nrw(hub = 'rawlanuv',query = 'pegel',descr = 'pegel_stand')
}

}
