% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R, R/utils.R
\name{list_page_zips}
\alias{list_page_zips}
\title{List ZIP file names from a web page}
\usage{
list_page_zips(page_url)

list_page_zips(page_url)
}
\arguments{
\item{page_url}{Character. URL of the webpage where ZIP files are listed.}
}
\value{
Character vector with ZIP file names.

Character vector of ZIP file names.
}
\description{
Loads an HTML page and extracts all ZIP file names from JSON-like structures.

This function fetches an HTML page and extracts the names of all linked
ZIP files. The output is a vector of file names.
}
\examples{
\dontrun{
list_page_zips("https://www.hochwasserportal.nrw/data/downloads/")
}
\dontrun{
list_page_zips("https://www.hochwasserportal.nrw/data/downloads/")
}
}
