% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_hydrologie_nrw.R
\name{load_filtered_station_data}
\alias{load_filtered_station_data}
\title{Load filtered station data from ZIP archives and combine them}
\usage{
load_filtered_station_data(page_url, file_index, startyear, endyear)
}
\arguments{
\item{page_url}{Character. Base URL of the page with ZIP files.}

\item{file_index}{data.frame with columns 'zip' and 'file'.}

\item{startyear}{Numeric. Start year.}

\item{endyear}{Numeric. End year.}
}
\value{
A combined `data.frame` or `NULL` if no files found.
}
\description{
Loads relevant files from ZIP archives for the specified year range,
reads CSVs, cleans values, and combines them into a single data frame.
}
